Minor: Split physical_plan/parquet/mod.rs into smaller modules
